refactor: Move get_initial_tip to zebra-state
