+++ zebra-consensus/src/chain/tests.rs
@@ -167,7 +167,7 @@ async fn verify_checkpoint() -> Result<(), Report> {
 
     // Test that the chain::init function works. Most of the other tests use
     // init_from_verifiers.
-    let mut chain_verifier = super::init(Mainnet, zebra_state::in_memory::init(), None);
+    let mut chain_verifier = super::init(Mainnet, zebra_state::in_memory::init()).await;
 
     /// SPANDOC: Make sure the verifier service is ready
     let ready_verifier_service = chain_verifier.ready_and().await.map_err(|e| eyre!(e))?;
