+++ zebra-state/Cargo.toml
@@ -9,19 +9,22 @@ edition = "2018"
 
 [dependencies]
 zebra-chain = { path = "../zebra-chain" }
-tower = "0.3.1"
-futures = "0.3.5"
-lazy_static = "1.4.0"
+
+color-eyre = "0.5"
 hex = "0.4.2"
-sled = "0.33.0"
+lazy_static = "1.4.0"
 serde = { version = "1", features = ["serde_derive"] }
+sled = "0.33.0"
+
+futures = "0.3.5"
+tower = "0.3.1"
 tracing = "0.1"
 tracing-futures = "0.2"
 
 [dev-dependencies]
-tokio = { version = "0.2.22", features = ["full"] }
 zebra-test = { path = "../zebra-test/" }
+
+once_cell = "1.4"
 spandoc = "0.2"
 tempdir = "0.3.7"
-color-eyre = "0.5"
-once_cell = "1.4"
+tokio = { version = "0.2.22", features = ["full"] }
